pyproject.toml
[tool.poetry]
name = "keeper"
version = "0.1.0"
description = "Trapper Keeper back-end"
authors = ["Jeff Hutchison <jeff@j3ff.io>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
djangorestframework = "^3.12.4"
psycopg2 = "^2.8.6"
python-dotenv = "^0.17.0"
django-extensions = "^3.1.3"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
black = "^21.4b2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
